pyproject.toml
[project]
name = "a2a-langgraph-fastapi-mini"
version = "0.1.0"
description = "Two LangGraph agents over A2A (JSON-RPC) with sync + streaming."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]==0.3.0",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "langgraph>=0.2.43",
    "langchain-core>=0.2.38",
    "langchain-openai>=0.2.2",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[project.scripts]
serve = "app.main:main"

[tool.uv]
# makes `uv run serve` available
package = true

[build-system]
requires = ["uv_build>=0.8.6,<0.9.0"]
build-backend = "uv_build"
